requirements.txt
fastapi
wtforms
starlette-wtf
pint
pyyaml
pydantic
python-multipart
minio
psycopg2
orator
pyjwt
requests
param
pandas
